% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\docType{data}
\name{AFIND_MAX_DISTANCE}
\alias{AFIND_MAX_DISTANCE}
\title{\enumerate{
\item limpar 0D 0D 0A para espaço
\item 
}

Ideia:
Aplicar o filtro verificador de instruções cloze
apenas no XML antes que este siga para o Moodle
evitando os erros lacónicos do Moodle.
Two problems
\enumerate{
\item se a linha tem mais que um multichoice
\item se uma mesma linha pode ter multichoice e numerical, etc
}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFIND_MAX_DISTANCE
}
\description{
One solution
scan, first, all multichoices, then all numerical, etc
}
\keyword{datasets}
