% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{cloze}
\alias{cloze}
\title{cloze - joins text and verify moodle-cloze {....} instructions}
\usage{
cloze(variant_title, variant_contents)
}
\arguments{
\item{variant_title}{a string}

\item{variant_contents}{a string}
}
\value{
alist with several attributes
}
\description{
cloze - joins text and verify moodle-cloze {....} instructions
}
