% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\name{cloze_transform}
\alias{cloze_transform}
\title{Transform text written in cloze format intro something else.}
\usage{
cloze_transform(cloze_text, output = "originalcloze")
}
\arguments{
\item{cloze_text}{text}

\item{output}{"I" (put "III") TODO...}
}
\value{
a list (cloze text parts)
}
\description{
Transform text written in cloze format intro something else.
}
