% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\docType{data}
\name{DBG_rmdexam}
\alias{DBG_rmdexam}
\title{TODO: Isto não funciona quando colocado no ficheiro .lintr:
\code{indentation_linter(indent = 2L, hanging_indent_style = c("tidy", "always", "never"), assignment_as_infix = TRUE)}
Ativa alguns \code{cat()} espalhados pelo código que
processa os exercícios. Este módulo faz parse
de texto: ativando esta flag passam a ser vistos
muitos \code{cat()} no processo de debug.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
DBG_rmdexam
}
\description{
TODO: Isto não funciona quando colocado no ficheiro .lintr:
\code{indentation_linter(indent = 2L, hanging_indent_style = c("tidy", "always", "never"), assignment_as_infix = TRUE)}
Ativa alguns \code{cat()} espalhados pelo código que
processa os exercícios. Este módulo faz parse
de texto: ativando esta flag passam a ser vistos
muitos \code{cat()} no processo de debug.
}
\keyword{datasets}
