% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{exer2rmdstring}
\alias{exer2rmdstring}
\title{exer2rmdstring}
\usage{
exer2rmdstring(nvar, ...)
}
\arguments{
\item{nvar}{number of variants (nrep <= VARCOUNT defined inside exercise)}

\item{...}{ex. "cap3/c3-estimativa.Rmd", "alinea01", "alinea04"}
}
\value{
string containing Rmd with # and ## sections (exercise title and variants)
}
\description{
exer2rmdstring
}
\examples{
   exerc2rmdstring(6, "cap3/c3-estimativa.Rmd", "alinea01", "alinea04")
}
