% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{extractquestions_fromhtml}
\alias{extractquestions_fromhtml}
\title{Extracts "moodle questions" from a html file}
\usage{
extractquestions_fromhtml(filename_no_extension)
}
\arguments{
\item{filename}{\itemize{
\item an user/author written filename
}}

\item{noneiscorrect}{}
}
\value{
\itemize{
\item a list with "moodle questions"
}
}
\description{
Extracts "moodle questions" from a html file
}
