% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{find_alinea}
\alias{find_alinea}
\title{Find "alinea" Rmd text by searching \code{alinea_tag}.}
\usage{
find_alinea(alinea_tag, ex_struct)
}
\arguments{
\item{alinea_tag}{}

\item{ex_struct}{see return value of \code{parse_exrmdfile()}}
}
\value{
string (rmd string)
}
\description{
\itemize{
\item \code{ex_struct$alineas}is a list of Rmd strings;
\item \code{ex_struct$alineas_title} is a list of titles (containing an \code{alinea_tag}).
}
}
