% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{find_setwd}
\alias{find_setwd}
\title{\code{find_setwd()} search for instructions \code{setwd()}}
\usage{
find_setwd(filename_no_extension)
}
\value{
nothing
}
\description{
The idea is to warn user to put data and images files in same path
or to add an absolut path.
}
\details{
Called in \code{xmlmoodle()}
}
