% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{get_ex_type}
\alias{get_ex_type}
\title{Get exercise type (cloze, essay, multichoice).}
\usage{
get_ex_type(curr_line)
}
\arguments{
\item{curr_line}{current line}
}
\value{
character
}
\description{
It is not working.
TODO: implement ex type.
}
