% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\name{get_exercise_root}
\alias{get_exercise_root}
\title{Get a global variable with the OS path of the questions
"database" in the filesystem .}
\usage{
get_exercise_root()
}
\value{
A string containing a OS path.
}
\description{
Get a global variable with the OS path of the questions
"database" in the filesystem .
}
