% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\name{maybe_pattern}
\alias{maybe_pattern}
\title{Before formal parse (with regex), it searches using \code{stringdist::afind}
for MULTICHOICE_S, NUMERICAL and SHORTANSWER patterns.
Matching is case-sensitive in \code{stringdist::afind}.}
\usage{
maybe_pattern(uclozetext)
}
\arguments{
\item{uclozetext}{\itemize{
\item an uppercase text
}}
}
\value{
FALSE or list(pattern="...pattern...", afind=afind_result)
}
\description{
Before formal parse (with regex), it searches using \code{stringdist::afind}
for MULTICHOICE_S, NUMERICAL and SHORTANSWER patterns.
Matching is case-sensitive in \code{stringdist::afind}.
}
