% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{next_line}
\alias{next_line}
\title{Increases \code{curr_lineno}}
\usage{
next_line(curr_lineno, all_lines)
}
\arguments{
\item{curr_lineno}{current line number}

\item{all_lines}{vector of lines of text}
}
\value{
integer
}
\description{
But it \code{cat()} if debug is activated.
}
