% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{parse_exrmdfile}
\alias{parse_exrmdfile}
\title{Parse an Rmd file containing an exercise ()}
\usage{
parse_exrmdfile(rmdfilename)
}
\arguments{
\item{rmdfilename}{}
}
\value{
list(ex_title, ex_type, code, enredo, alineas, alineas_title)
}
\description{
\code{parse_exrmdfile}
}
