% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\name{parse_one_cloze_instruction}
\alias{parse_one_cloze_instruction}
\title{Parse one cloze instruction}
\usage{
parse_one_cloze_instruction(cloze_instruction, output = "originalcloze")
}
\arguments{
\item{cloze_instruction}{\itemize{
\item instructions has here ...moodle cloze page....
}}

\item{output}{\itemize{
\item see line: if (output=="cloze")
}}
}
\value{
a list (cloze question parts)
}
\description{
Instructions are like {:SHORTANSWER:=Paris#Correct!~*#Wrong answer}
and breaks them into its parts.
}
\examples{
\dontrun{cloze_instruction <- "{:SHORTANSWER:=Paris#Correct!~*#Wrong answer}"}
\dontrun{print(parse_one_cloze_instruction(cloze_instruction, output="html"))}
\dontrun{cloze_instruction <- "{1:MCV:\%50\%Jupiter#Partial Correct!~=Mars#Correct~Saturn#Wrong}"}
\dontrun{print(parse_one_cloze_instruction(cloze_instruction, output="html"))}
}
