% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Trata-se de um "environment" local mas que permite mudar
o valor da variável "global" dentro deste package.}
\format{
An object of class \code{environment} of length 2.
}
\usage{
pkg_env
}
\description{
Trata-se de um "environment" local mas que permite mudar
o valor da variável "global" dentro deste package.
}
\keyword{datasets}
