% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{pq}
\alias{pq}
\title{\code{pq()} ("Planned Question") Extracts items from a given question. Since it's a planned
question, values inside its variants are accessible using \code{.[[VAR]]} or \code{.[VAR]}.}
\usage{
pq(...)
}
\arguments{
\item{...}{vector like "cap3/c3-estimativa.Rmd", "alinea01", "alinea04"}
}
\value{
string containing Rmd with # and ## sections (exercise title and variants)
}
\description{
\code{pq()} ("Planned Question") Extracts items from a given question. Since it's a planned
question, values inside its variants are accessible using \code{.[[VAR]]} or \code{.[VAR]}.
}
\examples{
\dontrun{planned_q("cap3/c3-estimativa.Rmd", "alinea01", "alinea04")}
}
