% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\name{remove_extension}
\alias{remove_extension}
\title{Remove an extension to a file name}
\usage{
remove_extension(filename, extension)
}
\arguments{
\item{filename}{possibly with or without extension}

\item{extension}{Rmd, R, .Rmd, .R, or any other}
}
\value{
character file name with extension
}
\description{
If the file name has the extension then
nothing is added.
}
\examples{
\dontrun{remove_extension("f",".Rmd")}
\dontrun{remove_extension("f",".R")}
\dontrun{remove_extension("f.Rmd","R")}
}
