% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{rmdexam}
\alias{rmdexam}
\title{Joins questions and items in an Rmd file (to be compiled).}
\usage{
rmdexam(rmdfilename, ...)
}
\arguments{
\item{rmdfilename}{filename to store the "rmd exam"}

\item{...}{each argument is a \code{c()} of strings;}
}
\value{
A file written in name given in \code{rmdfilename}
}
\description{
\code{rmdexam()} builds an Rmd file containing
an exam to be exported for moodle in a xml file.
After the revision of the author of the exam, then,
\code{xmlmoodle()} function must be called to make
the xml file. Importing to moodle has its own
"tricks" to be described elsewhere.
}
\examples{
\dontrun{rmdexam("my-new-moodle-exam.Rmd",
               pq("cap1/c1-estimation.Rmd", "est-mean", "est-var"),
               rq("cap2/c2-normalprob.Rmd", "prob-less", "prob-greater", "prob-between"))}
Remove 6 from rmdexam(6, "2022-2023-exame.Rmd",...)
TODO: maybe remove after all testing on new exams.
"arg" is like list(rmdtext="...", params=...)
Reproduce the instruction that
originates the exam for later
reproducibility.
Save exam in `rmdfilename`
Open and write
}
