% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{rq}
\alias{rq}
\title{\code{rq()} ("Random Question") Extracts items from a given question.}
\usage{
rq(...)
}
\arguments{
\item{...}{optional arguments and vector like "cap3/c3-estimativa.Rmd", "alinea01", "alinea04"}
}
\value{
string containing Rmd with # and ## sections (exercise title and variants)
}
\description{
Since it's a "random" question, code is replicated before each variant.
There is no need to use access \code{.[[VAR]]} or \code{.[VAR]} because there
is no pre calculated problems.
}
\details{
Optional arguments are:
\itemize{
\item seed (default a random integer): sets seed
\item varcount (default is 10): number of variants to be exported to moodle
}

A seed can help an author to revise an exam after it is verified or delivered
and corrected later.
}
\examples{
\dontrun{rq(seed = 90, number = 10, "cap3/c3-estimate.Rmd", "mean-item", "var-item")}
\dontrun{rq(number = 10, "cap3/c3-estimate.Rmd", "mean-item", "var-item")}
\dontrun{rq(seed = 90, "cap3/c3-estimate.Rmd", "mean-item", "var-item")}
\dontrun{rq("cap3/c3-estimate.Rmd", "mean-item", "var-item")}
}
