% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{saveto_xmlmoodle}
\alias{saveto_xmlmoodle}
\title{From a list of questions produce a xml file}
\usage{
saveto_xmlmoodle(
  filename_no_extension,
  exam_title,
  all_questions,
  em_fraction,
  defaultgrade
)
}
\arguments{
\item{exam_title}{\itemize{
\item name of the future file
}}

\item{all_questions}{\itemize{
\item a list
}}
}
\value{
\itemize{
\item a saved file in system
}
}
\description{
The xml file is to be imported in Moodle.
}
