% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdmoodle-package.R
\name{set_exercise_root}
\alias{set_exercise_root}
\title{Set a global variable with the OS path of the questions
"database" in the filesystem .}
\usage{
set_exercise_root(pathstr)
}
\arguments{
\item{pathstr}{string}
}
\value{
string
}
\description{
Set a global variable with the OS path of the questions
"database" in the filesystem .
}
\examples{
set_exercise_root("c:/Users/name/Documents/rmdmoodle")
}
