% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\name{set_warnings}
\alias{set_warnings}
\title{Turn TRUE or FALSE the
printing of warnings.}
\usage{
set_warnings(b)
}
\arguments{
\item{b}{\itemize{
\item a boolean TRUE or FALSE
}}
}
\value{
nothing
}
\description{
Turn TRUE or FALSE the
printing of warnings.
}
