% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{skip_chunk}
\alias{skip_chunk}
\title{To avoid search a pattern ^# Código or ^# enredo
etc, this function skips until a chunk ends.}
\usage{
skip_chunk(curr_lineno, all_lines, rmdfilename)
}
\value{
integer (line number after chunk ends)
}
\description{
To avoid search a pattern ^# Código or ^# enredo
etc, this function skips until a chunk ends.
}
