% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff.R
\name{slugify}
\alias{slugify}
\title{Transform "Title f(2024)" into "Title_f-2024-"}
\usage{
slugify(x, non_alphanum_replace = "-", space_replace = "_", tolower = FALSE)
}
\arguments{
\item{x}{character string}

\item{non_alphanum_replace}{"-" if nonalpha appears in \code{x}}

\item{space_replace}{"_" if spaces appears in \code{x}}

\item{tolower}{FALSE}
}
\value{
a string
}
\description{
Transform "Title f(2024)" into "Title_f-2024-"
}
\examples{
slugify("Title f(2024)")
}
