% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\name{validate_pattern}
\alias{validate_pattern}
\title{Formal verification of cloze instructions (using regex).}
\usage{
validate_pattern(uclozetext, resultofmaybe)
}
\arguments{
\item{uclozetext}{text with Moodle cloze instructions}

\item{resultofmaybe}{is list(pattern="...pattern...", afind=afind_result)}
}
\value{
list(type="...",  options = options, pos.s = pos.s,  pos.e = pos.e)
}
\description{
Formal verification of cloze instructions (using regex).
}
