% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{varcount_line}
\alias{varcount_line}
\title{Get VARCOUNT <- 6 or VARCOUNT = 6.
Search in the line for that pattern.}
\usage{
varcount_line(text_line)
}
\value{
numeric>0 or 0 if no varcount
}
\description{
Get VARCOUNT <- 6 or VARCOUNT = 6.
Search in the line for that pattern.
}
