% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifycloze.R
\name{verifycloze}
\alias{verifycloze}
\title{Check for wrong CLOZE instructions.}
\usage{
verifycloze(clozetext, showprogress = FALSE)
}
\arguments{
\item{sometext}{}
}
\value{
List of problems in output.
}
\description{
A ideia é fazer uma função \code{verifycloze(clozetext)} que
tente encontrar problemas em instruções cloze como
\itemize{
\item falta de  _S
\item falta de =
\item resultado como 8e-4 em numerical
i.e., se parece que é uma instrução cloze
então avisa caso ela não esteja com o formato desejado.
}
}
\details{
Depois do \code{knitr} é que se fica a saber
se existem situações de 8e-4, por exemplo.

Outra situação é com "~" que pode ser convertido em itálico:
\itemize{
\item procurar \if{html}{\out{<em>}} dentro de multichoice ou numerical ou etc.
}

Assim, esta rotina deve ser aplicada ao produto final em HTML.

Como informar o autor do local do erro
passando apenas um trecho do ficheiro!

Dificuldade:
\itemize{
\item em html perde-se a noção do contexto Rmd !!!
Solução:
}
}
\examples{
verifycloze("some text {:NUMERICAL:=10:0} etc\n etc\n {15:MULTICHOICE_S:=A~B~C}\n")

}
