% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{xmlmoodle}
\alias{xmlmoodle}
\title{Convert an Rmd file containing questions and variants
to xml moodle file to be imported for the "exam" tool
in moodle.}
\usage{
xmlmoodle(filename_no_extension)
}
\arguments{
\item{filename_no_extension}{}
}
\value{
A xml file to be imported in moodle.
}
\description{
Convert an Rmd file containing questions and variants
to xml moodle file to be imported for the "exam" tool
in moodle.
}
