% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{change_state}
\alias{change_state}
\title{Parser depends on states. This
function prints the next state if
it changes.}
\usage{
change_state(to, debug_str)
}
\arguments{
\item{to}{next state number}

\item{debug_str}{textual name of the next state number}
}
\value{
integer (a state is an integer)
}
\description{
Parser depends on states. This
function prints the next state if
it changes.
}
