% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{get_varcount}
\alias{get_varcount}
\title{Search for a pattern like "VARCOUNT <- 6"
in all lines (already read from rmdfile)}
\usage{
get_varcount(all_lines)
}
\arguments{
\item{all_lines}{}
}
\value{
integer (varcount value)
}
\description{
Search for a pattern like "VARCOUNT <- 6"
in all lines (already read from rmdfile)
}
