% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{parse_exrmdfile}
\alias{parse_exrmdfile}
\title{Parse an Rmd file containing an exercise}
\usage{
parse_exrmdfile(rmdfilename, varcount = NULL, number = 1)
}
\arguments{
\item{rmdfilename}{R Markdown file name.}

\item{varcount}{if null it extracts varcount from file in rmdfilename}

\item{number}{minimum 1}
}
\value{
list(ex_title, ex_type, code, enredo, alineas, alineas_title)
}
\description{
\code{varcount} is the number of variants of a question
from which a random one is selected to be presented to student.
}
\details{
\code{parse_exrmdfile}
}
\examples{
varcount is the number of
repetitions of a question
}
