% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdexam.R
\name{rmdexam}
\alias{rmdexam}
\title{Joins questions and items in an Rmd file (to be compiled).}
\usage{
rmdexam(nvariants, rmdfilename, ...)
}
\arguments{
\item{nvariants}{numeric (number of variants for each question)}

\item{rmdfilename}{filename to store the "rmd exam"}

\item{...}{each argument is a \code{c()} of strings;}
}
\value{
A file written in name given in \code{rmdfilename}
}
\description{
\code{rmdexam()} builds an Rmd file containing
an exam to be exported for moodle in a xml file.
After the revision of the author of the exam, then,
\code{xmlmoodle()} function must be called to make
the xml file. Importing to moodle has its own
"tricks" to be described elsewhere.
}
\examples{
make_rmdmoodle(6,
               "my-new-moodle-exam.Rmd",
               c("cap1/c1-estimation.Rmd","est-mean","est-var"),
               c("cap2/c2-normalprob.Rmd","prob-less", "prob-greater", "prob-between"))
}
