% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlmoodle.R
\name{xmlmoodle}
\alias{xmlmoodle}
\title{Exam in R Markdown to XML Moodle exam.}
\usage{
xmlmoodle(filename, params = NULL, em_fraction = NULL, defaultgrade = NULL)
}
\arguments{
\item{filename}{file where an exam in Rmd is stored}

\item{params}{list with same yaml params at top of R Markdown file}

\item{em_fraction}{string for multichoice questions ("-30", "-50", check Moodle because not all values are allowed)}

\item{defaultgrade}{string for multichoice questions  (1.0, 1.2, 2.5, 25, 50, etc, any positive value)}
}
\value{
A xml file to be imported in moodle.
}
\description{
Convert an Rmd file containing questions and variants
to xml moodle file to be imported for the "exam" tool
in moodle.
}
\examples{
\dontrun{xmlmoodle("my_exam.Rmd")}
\dontrun{xmlmoodle("my_exam")}
Warn author if `setwd()` is used.
Search for `setwd()` in order to warn
user to put data and images files in same path
or to add an absolut path.
Método sem multivariantes
TODO: o que são multivariantes ??

}
